### 2020-07-12 #2476 @whedon
[PRE REVIEW]: Disimpy: A massively parallel Monte Carlo simulator for synthesizing diffusion-weighted MRI data in Python
**Submitting author:** @kerkelae (<a href="http://orcid.org/0000-0001-9446-4305">Leevi Kerkelä</a>)
**Repository:** <a href="https://github.com/kerkelae/disimpy" target ="_blank">https://github.com/kerkelae/disimpy</a>
**Version:** v0.1
