### 2020-07-26 #2476 @Kevin-Mattheus-Moerman
@kerkelae can you work on those missing DOI's? You can call `@whedon generate pdf` to regenerate the paper and use `@whedon check references` to see if the DOI's are fixed.
