### 2020-07-26 #2476 @Kevin-Mattheus-Moerman
@mathieuboudreau, @grlee77, would you be interested in reviewing this work? You appear to have become my go-to-reviewers for MRI related stuff. Please do let me know if you feel I'm inviting you too often.

@ritagnunes @mfroeling this work may also interest you. Would you be able to help with this review too?

Thanks!
