### 2020-07-26 #2476 @Kevin-Mattheus-Moerman
@mstimberg @hanayik would you be interested in reviewing this work for JOSS?
The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and will focus on the [short paper](https://github.com/openjournals/joss-papers/blob/joss.02476/joss.02476/10.21105.joss.02476.pdf) and [the software](https://github.com/kerkelae/disimpy).

This is a pre-review issue where editors/reviewers are assigned. If you are interested in helping with this review you can reply here. I'll trigger a dedicated review issue once sufficient reviewers have been found.

Thanks.
