### 2020-07-26 #2476 @mathieuboudreau
What's the typical turnaround time you expect from reviewers? I'd be interested in reviewing, however my next three weeks or so are pretty packed because of the annual ISMRM (MRI) conference, and I am also already a reviewer for another JOSS submission.
