### 2020-07-27 #2476 @Kevin-Mattheus-Moerman
@mathieuboudreau thanks for getting back to me. I understand. We typically aim for 2 weeks but with all the COVID mayhem we've slowed down quite a bit. I hope to find 3 reviewers. I'll keep you on standby if that is okay and will add you if I cannot find sufficient reviewers. If so we will of course be patient to accommodate your time schedule.
