### 2020-07-27 #2476 @ritagnunes
Hi Kevin,

my schedule is also tight in the next few weeks with ISMRM, but as long as
it is not urgent, I am happy to review it too.

Thanks
Rita

On Sun, Jul 26, 2020 at 9:18 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @mathieuboudreau <https://github.com/mathieuboudreau>, @grlee77
> <https://github.com/grlee77>, would you be interested in reviewing this
> work? You appear to have become my go-to-reviewers for MRI related stuff.
> Please do let me know if you feel I'm inviting you too often.
>
> @ritagnunes <https://github.com/ritagnunes> @mfroeling
> <https://github.com/mfroeling> this work may also interest you. Would you
> be able to help with this review too?
>
> Thanks!
>
