+++ paper.bib
@@ -18,7 +18,8 @@
   booktitle={International Conference on Medical Image Computing and Computer-Assisted Intervention},
   pages={404--411},
   year={2010},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-15745-5_50}
 }
 
 @article{sjolund:2015,
@@ -28,7 +29,8 @@
   volume={261},
   pages={157--168},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jmr.2015.10.012}
 }
 
 @article{eriksson:2013,
@@ -38,7 +40,8 @@
   volume={226},
   pages={13--18},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jmr.2012.10.015}
 }
 
 @article{stejskal:1965,
@@ -49,7 +52,8 @@
   number={1},
   pages={288--292},
   year={1965},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1695690}
 }
 
 @article{blackman:2018,
@@ -64,7 +68,8 @@
   author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
+  doi={10.1145/2833157.2833162}
 }
 
 @article{virtanen:2020,
@@ -75,7 +80,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-019-0686-2}
 }
 
 @article{walt:2011,
@@ -86,7 +92,8 @@
   number={2},
   pages={22--30},
   year={2011},
-  publisher={IEEE Computer Society}
+  publisher={IEEE Computer Society},
+  doi={10.1109/mcse.2011.37}
 }
 
 @article{hall:2009,
@@ -97,7 +104,8 @@
   number={9},
   pages={1354--1364},
   year={2009},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tmi.2009.2015756}
 }
 
 @article{tournier:2007,
@@ -108,7 +116,8 @@
   number={4},
   pages={1459--1472},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2007.02.016}
 }
 
 @article{novikov:2019,
@@ -119,7 +128,8 @@
   number={4},
   pages={e3998},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/nbm.3998}
 }
 
 @article{assaf:2019,
@@ -130,7 +140,8 @@
   number={4},
   pages={e3762},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1101/140459}
 }
 
 @book{behrens:2009,
@@ -148,5 +159,6 @@
   number={2},
   pages={40--53},
   year={2008},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1109/hotchips.2008.7476516}
 }
