### 2020-08-05 #2527 @ritagnunes
@Kevin-Mattheus-Moerman, most likely I will only be able to work on this
after the end of the ISMRM conference, i.e. the week after next...

On Wed, Aug 5, 2020 at 11:40 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @DARSakthi <https://github.com/DARSakthi>, @grlee77
> <https://github.com/grlee77>, @ritagnunes <https://github.com/ritagnunes>
> this is just a friendly check in. I know some of you are busy on other
> reviews too. Could you give an update on when you expect to work on this
> review?
>
