### 2020-08-05 #2527 @Kevin-Mattheus-Moerman
Right, sorry I forgot about the ISMRM. Enjoy the conference!
