### 2020-08-05 #2527 @grlee77
@kerkelae, can you briefly describe the contributions of the coauthors your included? I could only verify your name directly in the commit history, but understand that there are other ways to contribute.
