### 2020-08-05 #2527 @grlee77
regarding the following review criteria:
> If there are any performance claims of the software, have they been confirmed?

Is the script that produced Fig. 1 available anywhere? I find the result believable, but did not try to install Camino to verify. I can confirm that all of the examples in the Jupyter notebook in the repository ran quickly on an NVIDIA GTX-1080 Ti card.
@Kevin-Mattheus-Moerman: Is this adequate verification for JOSS?
