### 2020-08-05 #2527 @grlee77
Overall the manuscript and software look good to me and I have only some minor suggestions.

**minor points to address**

1.) In the NumPy reference, the first author should be "van der Walt, S." rather than "Walt, S. van der"

2.) In the second sentence under features, I recommend the following grammatical change:
"generated in massively parallel on" -> "generated in a massively parallel fashion on"

3.) In the documentation's installation instructions a one-step command that might be easier for users than downloading the repository and then navigating to the folder to run pip would be to use:
```pip install git+https://github.com/kerkelae/disimpy.git```
(pip will take care of downloading the repository for the user)

4.) It may be worth providing a version of the demo notebook on [Google Colaboratory](https://colab.research.google.com/notebooks/intro.ipynb) as well, so users can try it out without having to own a CUDA-compatible GPU.

5.) Can the authors confirm if Camino's CPU-based Monte-Carlo is multi-threaded? If so, I would mention that explicitly in the manuscript text or Fig. 1 caption as it makes the relative improvement on the GPU more impressive.

6.) Regarding the "state of the filed" review criteria, I don't see much in the manuscript corresponding to this aside from the comparison to Camino. This is not an area I am particular familiar with the literature, so it would be good to know:
-  is Monte-Carlo is the dominant method of doing these simulations or are there other popular approaches?
- are there other popular tools aside from Camino for doing this?

7.) The included references look good, but I would include also reference previous work doing similar types of simulations on the GPU. For example:
```
Khieu-Van Nguyen, Edwin Hernández-Garzón, Julien Valette, Efficient GPU-based Monte-Carlo simulation of diffusion in real astrocytes reconstructed from confocal microscopy, Journal of Magnetic Resonance, Volume 296, 2018, Pages 188-199, https://doi.org/10.1016/j.jmr.2018.09.013.

Christopher A. Waudby, John Christodoulou, GPU accelerated Monte Carlo simulation of pulsed-field gradient NMR experiments, Journal of Magnetic Resonance, Volume 211, Issue 1, 2011, Pages 67-73, https://doi.org/10.1016/j.jmr.2011.04.004.
```
(optional)
It may also be worth mentioning other complementary software that could generate more realistic tissue models as input to the simulations. I am not familiar with specific tools for this, but am aware of the following publication (unfortunately the implementation described in it does not appear to be publicly available):

```
Kévin Ginsburger, Felix Matuschke, Fabrice Poupon, Jean-François Mangin, Markus Axer, Cyril Poupon,
MEDUSA: A GPU-based tool to create realistic phantoms of the brain microstructure using tiny spheres,
NeuroImage, Volume 193, 2019, Pages 10-24, https://doi.org/10.1016/j.neuroimage.2019.02.055
```

**misc comments**
Not having continuous integration seems okay for this library as the free CI services do not offer the needed GPU hardware.

Validation of accuracy of the simulations was not discussed in the manuscript, but from what I can tell, the authors have done validation via the following:
1.) examples in the tutorials:
        - free water diffusion matches the analytical solution
        - diffraction peaks generated as in Avram et. al. 2008
2.) via tests/test_simulations.py
        - some test cases compare to stored outputs of Camino
        - mesh-based sphere gives similar result to an analytical sphere
