### 2020-08-06 #2527 @fnery
Hi @grlee77, thanks for the feedback. Just a quick comment in terms of direct commits to the repo itself I didn't make a lot of them and they are no longer visible because we cleaned up the commit history for submission (made an orphan branch). They are still visible as [closed PRs](https://github.com/kerkelae/disimpy/pulls?q=is%3Apr+is%3Aclosed). In any case I will let @kerkelae clarify the describe authors' contributions.
