### 2020-08-06 #2527 @kerkelae
> @kerkelae, can you briefly describe the contributions of the coauthors you included? I could only verify your name directly in the commit history, but understand that there are other ways to contribute.

Good point @grlee77. I am the main author and have written most of the code and of the manuscript. Fabio Nery has helped by implementing various algorithms, making decisions on how to package the code, and testing. Matt Hall and Chris Clark have contributed to the software by being very involved in the planning and development phase of much of the code that the final package consists of.
