+++ paper.bib
@@ -86,7 +86,7 @@
 
 @article{walt:2011,
   title={The NumPy array: a structure for efficient numerical computation},
-  author={Walt, St{\'e}fan van der and Colbert, S Chris and Varoquaux, Gael},
+  author={Van der Walt, St{\'e}fan and Colbert, S Chris and Varoquaux, Gael},
   journal={Computing in science \& engineering},
   volume={13},
   number={2},
