+++ paper.bib
@@ -1,3 +1,57 @@
+@article{callaghan:2020,
+  title={ConFiG: Contextual Fibre Growth to generate realistic axonal packing for diffusion MRI simulation},
+  author={Callaghan, Ross and Alexander, Daniel C and Palombo, Marco and Zhang, Hui},
+  journal={Neuroimage},
+  volume={220},
+  pages={117107},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2020.117107}
+}
+
+@article{nguyen:2018,
+  title={Efficient GPU-based Monte-Carlo simulation of diffusion in real astrocytes reconstructed from confocal microscopy},
+  author={Nguyen, Khieu-Van and Hern{\'a}ndez-Garz{\'o}n, Edwin and Valette, Julien},
+  journal={Journal of Magnetic Resonance},
+  volume={296},
+  pages={188--199},
+  year={2018},
+  publisher={Elsevier},
+  doi={10.1016/j.jmr.2018.09.013}
+}
+
+@article{rafael:2020,
+  title={Robust Monte-Carlo Simulations in Diffusion-MRI: Effect of the substrate complexity and parameter choice on the reproducibility of results},
+  author={Rafael-Patino, Jonathan and Romascano, David and Ramirez-Manzanares, Alonso and Canales-Rodr{\'\i}guez, Erick Jorge and Girard, Gabriel and Thiran, Jean-Philippe},
+  journal={Frontiers in Neuroinformatics},
+  volume={14},
+  pages={8},
+  year={2020},
+  publisher={Frontiers},
+  doi={10.3389/fninf.2020.00008}
+}
+
+@article{li:2019,
+  title={SpinDoctor: a Matlab toolbox for diffusion MRI simulation},
+  author={Li, Jing-Rebecca and Nguyen, Van-Dang and Tran, Try Nguyen and Valdman, Jan and Trang, Cong-Bang and Van Nguyen, Khieu and Vu, Duc Thach Son and Tran, Hoang An and Tran, Hoang Trong An and Nguyen, Thi Minh Phuong},
+  journal={NeuroImage},
+  volume={202},
+  pages={116120},
+  year={2019},
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2019.116120}
+}
+
+@inproceedings{ianus:2016,
+  title={Microstructure imaging sequence simulation toolbox},
+  author={Ianu{\c{s}}, Andrada and Alexander, Daniel C and Drobnjak, Ivana},
+  booktitle={International Workshop on Simulation and Synthesis in Medical Imaging},
+  pages={34--44},
+  year={2016},
+  organization={Springer},
+  doi={10.1007/978-3-319-46630-9_4}
+}
+
 @inproceedings{gyori:2020,
   title={Discrepancy between in-vivo measurements and Monte-Carlo simulations with spherical structures in B-tensor encoding},
   author={Gyori, Noemi and Hall, Matt G and Clark, Chris A and Alexander, Daniel C and Kaden, Enrico},
