### 2020-08-06 #2527 @DARSakthi
Overall the paper is well done. The tutorial section of the documentation is a nice touch, and the mathematical specification in the final section of the paper is also appreciated. Having a colab notebook would be an excellent idea.

I echo the suggestions of @grlee77, in particular verification of the performance claims. I do question the use of the word synthesising in the title -- while 'to synthesise' can mean to produce artificially, a more common usage is to combine. If the authors are partial to the title then no change is necessary, but a better word may clarify the function of the software.
