### 2020-08-06 255c75a
typo fix and small addition
