### 2020-08-06 e2a10c5
simplified install instructions
