+++ docs/source/installation.rst
@@ -38,16 +38,14 @@ Installation
 
 To avoid possible issues with automatic installation of some of the
 dependencies, we recommend installing Disimpy in a virtual environment or a
-conda environment. Disimpy can be installed by downloading the source code from
-`here <https://github.com/kerkelae/disimpy/archive/v0.1.zip>`_. Extract the
-downloaded .zip file, navigate into the extracted directory, and install
-with pip: ::
+conda environment. Disimpy can be installed with pip by executing: ::
 
-    pip install .
+    pip install https://github.com/kerkelae/disimpy/archive/v0.1.tar.gz
 
 Alternatively, given that the requirements specified above are met, you can use
 Disimpy without installing by directly importing the functions you want to use
-from the extracted directory.
+from the directory you can download
+`here <https://github.com/kerkelae/disimpy/archive/v0.1.zip>`_.
 
 Automated tests
 ###############
