### 2020-08-11 272d29d
instructions to install latest master
