+++ paper.md
@@ -54,7 +54,7 @@ powerful tool for investigating the accuracy and precision of new methods, e.g.,
 [@tournier:2007]. Generally, dMRI simulations are based on modelling diffusion
 inside some geometry to obtain a solution to the diffusion equation, e.g.,
 [@li:2019; @ianus:2016], or modelling diffusion using a more generalizable Monte
-Carlo approach, e.g., [@hall:2009 @rafael:2020]. The Monte Carlo method enables
+Carlo approach, e.g., [@hall:2009; @rafael:2020]. The Monte Carlo method enables
 the use of complex and realistic tissue microstructure models, e.g.,
 [@callaghan:2020], and can be significantly accelerated using GPU computing,
 e.g., [@nguyen:2018].
