### 2020-08-12 #2527 @kerkelae
@grlee77 and @DARSakthi, thank you for the feedback. We have made a number of changes to the manuscript and the documentation. Please see the points below.

- We made the small fixes and changes suggested by @grlee77. We also have no particular preference for the word "synthesize" which we replaced by "generate". The updated manuscript contains the changed title but I am not sure how to change the title of my JOSS submission. Could @Kevin-Mattheus-Moerman help us with this please?

- Please find [here](https://github.com/openjournals/joss-reviews/files/5061338/simple_comparison.zip) a simple script and a ply file that can be used to reproduce Figure 1. Since we didn't find large differences between runtimes of individual simulations, we reported a single set of runtimes instead of averages. Please note that reproducing the figure requires the installation of Java and Camino. Camino is single-threaded and many studies using Camino have run simulations in parallel on clusters.

- We simplified the installation instructions according to the suggestion by @grlee77 and expanded the tutorial to describe how to run the examples on Google Colaboratory. Special thanks for this tip!

- We expanded the "Statement of need" section to answer the points 6 and 7 by @grlee77. There exists other methods for simulating dMRI experiments, for example, by solving the Bloch-Torrey equation (which adds the diffusion term to the Bloch equation), but the Monte Carlo method is more generalizable. To our knowledge, Camino is the most popular Monte Carlo simulator but other simulators have been recently developed, some of which are GPU-accelerated. However, we are not aware of approachable Python packages for performing GPU-accelerated dMRI simulations which we find to be of value. We also mentioned the need for relevant microstructural models.

- Indeed, we do not have continuous integration since the services we are familiar with do not provide support for CUDA. To validate the functionality of the package, we have written unit tests and compared the simulated signals to analytical solutions and to Camino's results.
