### 2020-08-12 #2527 @DARSakthi
Is the test script you linked to available in the repo too? I can't seem to find it. If not, I think it could be a good thing to include from an end - user stand point, so as to evaluate whether to use the software or not.

With the most recent round of changes, I am happy to recommend the paper for publication. Well done!
