### 2020-08-17 #2527 @ritagnunes
I really enjoyed being able to run the Tutorial on Google Colaboratory. That was a really nice addition!
It will be very helpful for potential/new users to have a quick grasp of the implemented software features.

Also, the provided examples make it easy to verify the requirement of simulating a high number of random walkers to match the theoretical predictions motivating the need for this software.  I also like it that the authors have included a range of microstructural environments, from free to restricted diffusion including multiple compartments and a diffraction example replicating a previous study.

It was harder for me to evaluate the gains in simulation speed as the software and corresponding tutorial do not run without GPU-enabled acceleration. Although I did not run Camino independently, I trust that the simulation times reported in the tests are correct.

The provided documentation is quite comprehensive and the paper clearly motivates the need for the software, describing the implemented features. I also really like that the basic theoretical equations have been included.

I agree with @DARSakthi and I am also happy to recommend the publication of this paper.
Congratulations!
