### 2020-08-25 #2527 @Kevin-Mattheus-Moerman
@openjournals/dev can you help advise as to how to change the title of a submission?
