### 2020-08-25 #2527 @arfon
> @openjournals/dev can you help advise as to how to change the title of a submission?

Just change the title of the paper in the `paper.md` file.
