### 2020-08-25 #2527 @kerkelae
Thank you to all the reviewers!

@arfon and @Kevin-Mattheus-Moerman, I have changed the title in the `paper.md` file. I just wanted to know if there is something more I need to do because the old title was still shown on the [JOSS website](https://joss.theoj.org/papers/ee3aa136859cb47595de7de8eb84b441)
