### 2020-08-25 #2527 @arfon
> Thanks @arfon and @Kevin-Mattheus-Moerman. I had changed the title in the paper.md file and just wanted to know if there is something more I need to do because the old title was still shown on the JOSS website.

The version on the JOSS site will update when the paper is published.
