### 2020-08-26 #2527 @Kevin-Mattheus-Moerman
@kerkelae I've checked your paper and can confirm it looks in order.

These are final steps to complete:

- [ ] Please archive a copy of the software on ZENODO and report back here with the DOI of the archived version. When you create this archived version be sure to have the title and author meta data match that of the paper (you may need to manually alter this).

- [ ] Can you inform me of the most recent version tag for this work? Is it still at `v0.1`?
