### 2020-08-26 #2527 @kerkelae
@Kevin-Mattheus-Moerman

Since we changed a few things during the review, I created a [new release](https://github.com/kerkelae/disimpy/releases/tag/v0.1.1): `v0.1.1`. I archived this version on [Zenodo](https://zenodo.org/record/4001687). The DOI is http://doi.org/10.5281/zenodo.4001687.
