### 2020-08-26 bed14cb
added reference to JOSS paper
