+++ README.rst
@@ -4,11 +4,9 @@ Disimpy
 
 Disimpy is a tool for generating synthetic diffusion magnetic resonance imaging
 data that is useful in the development and validation of new data acquisition
-and analysis methods.
-
-The synthetic data is generated by massively parallelized Monte Carlo
-simulations on Nvidia's CUDA-capable GPUs according to the simulation framework
-established by Hall et al. [1]_.
+and analysis methods. The synthetic data is generated using massively
+parallelized Monte Carlo simulations on Nvidia's CUDA-capable GPUs. Details
+of the signal generation can be found in [1]_.
 
 Requirements and installation
 #############################
@@ -33,7 +31,6 @@ about problems or bugs, please read the `contributing guidelines
 References
 ==========
 
-.. [1] Hall, Matt G., and Daniel C. Alexander. "Convergence and parameter choice
-       for Monte-Carlo simulations of diffusion MRI." IEEE transactions on
-       medical imaging 28.9 (2009): 1354-1364. doi:`10.1109/TMI.2009.2015756
-       <https://ieeexplore.ieee.org/document/4797853>`_
+.. [1] Kerkelä et al., (2020). Disimpy: A massively parallel Monte Carlo
+       simulator for generating diffusion-weighted MRI data in Python. Journal
+       of Open Source Software, 5(52), 2527.https://doi.org/10.21105/joss.02527
